inheritance test also added for other classes
